+++ .idea/misc.xml
@@ -6,17 +6,23 @@
         <entry key="..\:/AppMisionTIC/app/src/main/res/drawable-v23/splash.xml" value="0.3828125" />
         <entry key="..\:/AppMisionTIC/app/src/main/res/drawable/degrade.xml" value="0.3828125" />
         <entry key="..\:/AppMisionTIC/app/src/main/res/drawable/splash.xml" value="0.3828125" />
+        <entry key="..\:/AppMisionTIC/app/src/main/res/layout/activity_main.xml" value="0.20416666666666666" />
+        <entry key="..\:/AppMisionTIC/app/src/main/res/layout/activity_main2.xml" value="0.15896739130434784" />
         <entry key="..\:/AppMisionTIC/app/src/main/res/layout/activity_splash.xml" value="0.35260416666666666" />
-        <entry key="..\:/Users/COMPUMAX/AndroidStudioProjects/AppMisionTIC/app/src/main/res/drawable-v23/splash.xml" value="0.3769230769230769" />
-        <entry key="..\:/Users/COMPUMAX/AndroidStudioProjects/AppMisionTIC/app/src/main/res/drawable/degrade.xml" value="0.3769230769230769" />
+        <entry key="..\:/AppMisionTIC/app/src/main/res/layout/card_layout.xml" value="0.1" />
+        <entry key="..\:/AppMisionTIC/app/src/main/res/layout/detail_fragment.xml" value="0.17255434782608695" />
+        <entry key="..\:/AppMisionTIC/app/src/main/res/layout/fragment_config_.xml" value="0.16666666666666666" />
+        <entry key="..\:/AppMisionTIC/app/src/main/res/layout/fragment_list_.xml" value="0.1" />
+        <entry key="..\:/AppMisionTIC/app/src/main/res/layout/fragment_menu_.xml" value="0.15896739130434784" />
         <entry key="..\:/Users/COMPUMAX/AndroidStudioProjects/AppMisionTIC/app/src/main/res/drawable/icon_back.xml" value="0.3828125" />
         <entry key="..\:/Users/COMPUMAX/AndroidStudioProjects/AppMisionTIC/app/src/main/res/drawable/logo_turistic3_background.xml" value="0.3828125" />
         <entry key="..\:/Users/COMPUMAX/AndroidStudioProjects/AppMisionTIC/app/src/main/res/drawable/menu_puntos.xml" value="0.3828125" />
         <entry key="..\:/Users/COMPUMAX/AndroidStudioProjects/AppMisionTIC/app/src/main/res/drawable/menu_puntos_negro.xml" value="0.3828125" />
-        <entry key="..\:/Users/COMPUMAX/AndroidStudioProjects/AppMisionTIC/app/src/main/res/drawable/splash.xml" value="0.3769230769230769" />
         <entry key="..\:/Users/COMPUMAX/AndroidStudioProjects/AppMisionTIC/app/src/main/res/layout/activity_main.xml" value="0.32592592592592595" />
         <entry key="..\:/Users/COMPUMAX/AndroidStudioProjects/AppMisionTIC/app/src/main/res/layout/activity_main2.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/COMPUMAX/AndroidStudioProjects/AppMisionTIC/app/src/main/res/layout/card_layout.xml" value="0.36666666666666664" />
+        <entry key="..\:/Users/COMPUMAX/AndroidStudioProjects/AppMisionTIC/app/src/main/res/layout/detail_fragment.xml" value="0.36666666666666664" />
+        <entry key="..\:/Users/COMPUMAX/AndroidStudioProjects/AppMisionTIC/app/src/main/res/layout/fragment_list_.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/COMPUMAX/AndroidStudioProjects/AppMisionTIC/app/src/main/res/mipmap-anydpi-v26/logo_turistic3.xml" value="0.3828125" />
         <entry key="..\:/Users/COMPUMAX/Desktop/AppMisionTIC/app/src/main/res/layout/activity_main.xml" value="0.36666666666666664" />
         <entry key="..\:/Users/COMPUMAX/Downloads/AppMisionTIC/app/src/main/res/layout/activity_main.xml" value="0.36666666666666664" />
