+++ app/src/main/java/com/misiontic/touristicsites/SplashActivity.kt
@@ -7,6 +7,6 @@ import android.os.Bundle
 class SplashActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        startActivity(Intent(this,MainActivity2::class.java))
+        startActivity(Intent(this,MainActivity::class.java))
     }
 }
