+++ app/src/main/res/drawable-v23/splash.xml
@@ -6,9 +6,7 @@
         android:height="240dp"
         android:left="60dp"
         android:right="60dp"
-        android:top="240dp"
-        android:bottom="240dp">
-
+        android:top="240dp">
         <bitmap android:src="@drawable/logo_turistic2"/>
     </item>
 </layer-list>
