+++ app/src/main/res/drawable/splash.xml
@@ -2,10 +2,11 @@
 <layer-list xmlns:android="http://schemas.android.com/apk/res/android">
     <item android:drawable="@drawable/degrade"/>
     <item
-        android:width="100dp"
-        android:height="100dp"
+        android:width="240dp"
+        android:height="240dp"
         android:left="90dp"
-        android:top="200dp">
+        android:right="60dp"
+        android:top="240dp">
         <bitmap android:src="@drawable/logo_turistic2"/>
     </item>
 </layer-list>
