+++ app/src/main/res/layout/activity_main.xml
@@ -4,160 +4,15 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/black"
     tools:context=".MainActivity">
 
-    <ImageView
-        android:id="@+id/iv_logo"
-        android:layout_width="150dp"
-        android:layout_height="50dp"
-        android:layout_margin="10dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/logo_turistic_horizontal" />
+   <fragment
+       android:name="com.misiontic.touristicsites.List_Fragment"
+       android:layout_width="match_parent"
+       android:layout_height="match_parent"
+       android:id="@+id/fragment_recycler">
 
-    <ImageButton
-        android:layout_width="wrap_content"
-        android:layout_height="60dp"
-        android:layout_marginStart="16dp"
-        android:background="@color/black"
-        android:src="@drawable/icon_back"
-        app:layout_constraintBottom_toTopOf="@id/tvDubai"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+   </fragment>
 
-    <ImageButton
-        android:layout_width="wrap_content"
-        android:layout_height="60dp"
-        android:layout_marginEnd="16dp"
-        android:background="@color/black"
-        android:src="@drawable/menu_puntos"
-        app:layout_constraintBottom_toTopOf="@id/tvDubai"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/tvDubai"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
-        android:background="@color/teal_200"
-        android:gravity="center"
-        android:padding="5dp"
-        android:textColor="@color/black"
-        android:textSize="24sp"
-        android:textStyle="bold"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/iv_logo" />
-
-    <ImageView
-        android:id="@+id/iv_dubai_1"
-        android:layout_width="match_parent"
-        android:layout_height="200dp"
-        android:scaleType="centerCrop"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/tvDubai"
-        app:srcCompat="@drawable/dubai_1" />
-
-    <TextView
-        android:id="@+id/tv_Subtitulo1"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="12dp"
-        android:text="@string/subtitulo_descripcion"
-        android:textColor="@color/teal_200"
-        android:textStyle="bold"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/iv_dubai_1" />
-
-    <TextView
-        android:id="@+id/tv_detalle"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="12dp"
-        android:justificationMode="inter_word"
-        android:textColor="@color/white"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/tv_Subtitulo1" />
-
-    <TextView
-        android:id="@+id/tv_Subtitulo2"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="12dp"
-        android:text="@string/subtitulo_ubicacion"
-        android:textColor="@color/teal_200"
-        android:textStyle="bold"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/tv_detalle" />
-
-    <TextView
-        android:id="@+id/tv_detalle2"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="12dp"
-        android:justificationMode="inter_word"
-        android:textColor="@color/white"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/tv_Subtitulo2" />
-
-    <TextView
-        android:id="@+id/tv_Subtitulo3"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="12dp"
-        android:layout_marginStart="72dp"
-        android:layout_marginTop="136dp"
-        android:text="@string/subtitulo_tiempo"
-        android:textColor="@color/teal_200"
-        android:textStyle="bold"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/tv_detalle2" />
-
-    <TextView
-        android:id="@+id/tv_detalle3"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="12dp"
-        android:justificationMode="inter_word"
-        android:textColor="@color/white"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/tv_Subtitulo3" />
-
-    <TextView
-        android:id="@+id/tv_Subtitulo4"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="12dp"
-        android:text="@string/subtitulo_sitios"
-        android:textColor="@color/teal_200"
-        android:textStyle="bold"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/tv_detalle3" />
-
-    <TextView
-        android:id="@+id/sitios"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="12dp"
-        android:justificationMode="inter_word"
-        android:textColor="@color/white"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/tv_Subtitulo4" />
-
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="20dp"
-        android:background="@color/teal_200"
-        android:gravity="center"
-        android:text="@string/pie_de_pagina"
-        android:textColor="@color/black"
-        android:textSize="12dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
