+++ app/src/main/res/navigation/navigation.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/navigation"
+    app:startDestination="@id/list_Fragment">
+
+    <fragment
+        android:id="@+id/list_Fragment"
+        android:name="com.misiontic.touristicsites.List_Fragment"
+        android:label="fragment_list_"
+        tools:layout="@layout/fragment_list_" />
+    <fragment
+        android:id="@+id/config_Fragment"
+        android:name="com.misiontic.touristicsites.Config_Fragment"
+        android:label="fragment_config_"
+        tools:layout="@layout/fragment_config_" />
+</navigation>
